Major refactor for Input part. Added failure transparency, file validation, regex for urls, etc.
